+++ src/main/java/com/demospringfullstack/springbootexample/customer/CustomerDTO.java
@@ -3,7 +3,7 @@ package com.demospringfullstack.springbootexample.customer;
 import java.util.List;
 
 public record CustomerDTO(
-        Integer id,
+        Long id,
         String name,
         String email,
         Gender gender,
