+++ src/main/java/com/demospringfullstack/springbootexample/customer/CustomerJPADataAccessService.java
@@ -18,7 +18,7 @@ public class CustomerJPADataAccessService implements CustomerDAO {
     }
 
     @Override
-    public Optional<Customer> selectCustomerById(Integer id) {
+    public Optional<Customer> selectCustomerById(Long id) {
 
         return customerRepository.findById(id);
     }
@@ -35,13 +35,13 @@ public class CustomerJPADataAccessService implements CustomerDAO {
     }
 
     @Override
-    public boolean existsPersonWithId(Integer id) {
+    public boolean existsPersonWithId(Long id) {
 
         return customerRepository.existsCustomerById(id);
     }
 
     @Override
-    public void removeCustomerById(Integer id) {
+    public void removeCustomerById(Long id) {
         customerRepository.deleteById(id);
     }
